requirements.txt
python-telegram-bot
matplotlib
numpy
statsmodels